+++ examples/test/case/test_case_range_badbackwards.pas
@@ -1,0 +1,8 @@
+program TestCaseRangeBadBackwards;
+var
+    x: integer;
+begin
+    case x of
+        5..1: writeln('a');
+    end;
+end.