+++ examples/test/case/test_case_range_badoverlap.pas
@@ -1,0 +1,9 @@
+program TestCaseRangeBadOverlap;
+var
+    x: integer;
+begin
+    case x of
+        1..5: writeln('a');
+        3..7: writeln('b');
+    end;
+end.