+++ examples/test/case/test_case_range_badtypemismatch.pas
@@ -1,0 +1,8 @@
+program TestCaseRangeBadTypeMismatch;
+var
+    x: integer;
+begin
+    case x of
+        1..'z': writeln('a');
+    end;
+end.