+++ examples/test/case/test_case_range_badvalueinrange.pas
@@ -1,0 +1,9 @@
+program TestCaseRangeBadValInRange;
+var
+    x: integer;
+begin
+    case x of
+        1..5: writeln('a');
+        3: writeln('b');
+    end;
+end.