+++ examples/test/case/test_case_range_char.pas
@@ -1,0 +1,28 @@
+program TestCaseRangeChar;
+var
+    c: char;
+begin
+    c := 'a';
+    case c of
+        'a'..'z': writeln('lower');
+        'A'..'Z': writeln('upper');
+    else
+        writeln('other');
+    end;
+
+    c := 'Z';
+    case c of
+        'a'..'z': writeln('lower');
+        'A'..'Z': writeln('upper');
+    else
+        writeln('other');
+    end;
+
+    c := '5';
+    case c of
+        'a'..'z': writeln('lower');
+        'A'..'Z': writeln('upper');
+    else
+        writeln('other');
+    end;
+end.